fixed some dataTransfer bugs

+++ app/src/main/java/xyz/farshad/vocab/component/Server/DataTransfer.java
@@ -1,17 +1,27 @@
 package xyz.farshad.vocab.component.Server;
 
+import android.content.Context;
 import android.os.AsyncTask;
+import android.util.Log;
+import android.widget.Toast;
 
 import org.json.JSONException;
 import org.json.JSONObject;
 
-import xyz.farshad.vocab.component.DataSource.Import;
+import xyz.farshad.vocab.component.DataService.Import;
 import xyz.farshad.vocab.model.Course;
 import xyz.farshad.vocab.model.Level;
 import xyz.farshad.vocab.model.Word;
 
 public class DataTransfer extends AsyncTask {
+
+    private Context context;
     private RemoteServer server;
+
+    public DataTransfer(Context context) {
+        this.context = context;
+    }
+
     @Override
     protected void onPreExecute() {
         super.onPreExecute();
@@ -21,33 +31,35 @@ public class DataTransfer extends AsyncTask {
     protected Object doInBackground(Object[] objects) {
         String link = "http://192.168.0.101:8080/api/retrieve";
         server = new RemoteServer();
-        try{
-            JSONObject responseData = server.get(link);
-            return responseData;
-        }catch(Exception e){
-            return new String("can not connect to server: " + e.getMessage());
-        }
+        Object responseData = server.get(link);
+        return responseData;
     }
 
     @Override
     protected void onPostExecute(Object o) {
         super.onPostExecute(o);
 
-        //delete all records
-        Course.deleteAll(Course.class);
-        Level.deleteAll(Level.class);
-        Word.deleteAll(Word.class);
+         if (o instanceof Exception){
+             Log.i("WEB", ((Exception) o).getMessage());
+             Toast.makeText(context, "Can not connect to server!", Toast.LENGTH_LONG).show();
+         }else {
+             //delete all records
+             Course.deleteAll(Course.class);
+             Level.deleteAll(Level.class);
+             Word.deleteAll(Word.class);
 
-        JSONObject responseData = (JSONObject) o;
+             JSONObject responseData = (JSONObject) o;
 
-        //import all records
-        Import imp = new Import();
-        try {
-            imp.importCourse(responseData.getJSONArray("course"));
-            imp.importLevel(responseData.getJSONArray("level"));
-            imp.importWord(responseData.getJSONArray("word"));
-        } catch (JSONException e) {
-            e.printStackTrace();
-        }
+             //import all records
+             Import imp = new Import();
+             try {
+                 imp.importCourse(responseData.getJSONArray("course"));
+                 imp.importLevel(responseData.getJSONArray("level"));
+                 imp.importWord(responseData.getJSONArray("word"));
+             } catch (JSONException e) {
+                 e.printStackTrace();
+             }
+         }
+
     }
 }
